+++ src-tauri/src/commands/auth/auth_methods.rs
@@ -1,6 +1,5 @@
 use log::info;
 use tauri::State;
-use zeroize::Zeroize;
 
 use super::DiaryState;